+++ tasks/PowerShell-Matrix.Tasks.ps1
@@ -10,23 +10,29 @@ task UpdateBuildTasks {
 #region GitHub Actions
 
 task UpdateValidationWorkflow {
+    [System.IO.FileInfo] $file = "$PSScriptRoot/../.github/workflows/build-validation.yml"
+    New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
         -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/build-validation-matrix.yml' `
-        -OutFile "$PSScriptRoot\..\.github\workflows\build-validation.yml"
+        -OutFile $file
 }
 
 task UpdatePreReleaseWorkflow {
+    [System.IO.FileInfo] $file = "$PSScriptRoot\..\.github\workflows\pre-release.yml"
+    New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
         -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/pre-release-windows.yml' |
     ForEach-Object { $_ -replace 'MyModuleName', $ModuleName } |
-    Out-File "$PSScriptRoot\..\.github\workflows\pre-release.yml"
+    Out-File $file -NoNewline
 }
 
 task UpdateReleaseWorkflow {
+    [System.IO.FileInfo] $file = "$PSScriptRoot\..\.github\workflows\release.yml"
+    New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
         -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/release-windows.yml' |
     ForEach-Object { $_ -replace 'MyModuleName', $ModuleName } |
-    Out-File "$PSScriptRoot\..\.github\workflows\release.yml"
+    Out-File $file -NoNewline
 }
 
 #endregion