+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Dotnet-Matrix.Tasks does not fail in empty project.
+- PowerShell-Matrix.Tasks does not fail in empty project.
 - Wrong file separators for linux.
 
 ## [1.1.0] - 2022-03-19